Refactor client side db functions
