+++ lib/db.js
@@ -1,31 +1,22 @@
 import { doc, deleteDoc, setDoc, serverTimestamp } from 'firebase/firestore';
 
 import { db } from './firebase';
-import { formatCapitalize } from '../utils/utils';
+import { formatCapitalize } from '../utils';
 
-export const handleCreateUser = async (user, providerId, email) => {
-  try {
-    await setDoc(
-      doc(db, 'users', user.uid),
-      {
-        name: user?.displayName,
-        email: user?.email ?? email,
-        providerId:
-          formatCapitalize(providerId?.split('.')[0]) ?? 'Email and password',
-        timestamp: serverTimestamp(),
-      },
-      { merge: true }
-    );
-  } catch (error) {
-    console.log(error);
-  }
+export const createFirestoreUser = (user, providerId, email) => {
+  return setDoc(
+    doc(db, 'users', user.uid),
+    {
+      name: user?.displayName,
+      email: user?.email ?? email,
+      providerId:
+        formatCapitalize(providerId?.split('.')[0]) ?? 'Email and password',
+      timestamp: serverTimestamp(),
+    },
+    { merge: true }
+  );
 };
 
-export const handleDeleteUser = async (uid) => {
-  // TODO only deletes user doc, not the subcollection of invoices. create cloud function to handle deletes
-  try {
-    await deleteDoc(doc(db, 'users', uid));
-  } catch (error) {
-    console.log(error);
-  }
+export const deleteFirestoreUser = (uid) => {
+  return deleteDoc(doc(db, 'users', uid));
 };
